requirements.txt
torch~=1.13.0
modules~=1.0.0
opencv-python~=4.5.5.64
numpy~=1.22.4
pandas~=1.4.2
tqdm~=4.64.0
matplotlib~=3.5.2
Pillow~=9.1.1
scipy~=1.8.1
requests~=2.28.1
beautifulsoup4~=4.11.1
scenedetect~=0.6.0.3
pre-commit
seaborn~=0.11.2
statsmodels~=0.13.2
Flag~=0.1.1
